Add the option to fix solar hour (solar angle) and remove the old and incomplete options.
This change aims to add an option (fixed_solhr) to fix solar hour (solar angle) that is similar with fixed_date. This option fixes the solar hour to the initial time. It is useful for doubly-periodic limited domain simulation that may not want the diurnal cycle. Meanwhile, remove two out-of-date options (fix_cosz_dec and fix_cosz_shr) that cannot fully remove diurnal cycle.