Add an option to calculate daily mean solar zenith angle.
This option (daily_mean) aims to replace the hourly mean solar zenith angle with the daily mean value. This is useful for long time (>> 1 day) doubly-periodic limited domain simulation.

Without this option, when the solar angle is fixed, the sun will stay at a fixed location where there could be too much solar heating. If the model runs for a few hours or less, that is fine. But, if the model runs for more than one day, the solar heating will be too much under the sun and too little in the dark.