Shallow convection modification (mfshalcnv.f)
This change is to address the unrealistic heating tendency at the arbitrarily enforced cloud top (P/Ps=0.7 or ~700mb) in the original scheme. The issue is particularly obvious when deep convection is turned off. The main change is that the cloud depth and top is checked before calling the shallow convection scheme. If the cloud is too deep or too high, it should not be handled by shallow convection.

Specific changes:
- reset kmax as the new cloud top limiter (if ktcon > kmax, shal conv
not called)
- add critical cloud depth and height (cthk_shal and top_shal) as namelist parameters
- register updraft velocity square (wu2) and normalized mass flux (eta)
as 3D outputs