Add an namelist option scale_awareness_factor.
This namelist option is designed to adjust the scale-awareness of the
convection and PBL schemes. Our findings indicate that when the model
resolution changes from 13 km to 6.5 km, there is a noticeable
degradation in large-scale circulation and tropical cyclone (TC) track
prediction. This suggests that the scale-awareness in the convection and
PBL schemes is not optimal for resolutions other than 13 km. The default
value of scale_awareness_factor is set to 1.0. For a resolution of 6.5
km, I recommend increasing the scale_awareness_factor to 2.0.